requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytesseract
motor
pydantic
passlib[bcrypt]
pyjwt
rapidfuzz
numpy
opencv-python-headless
python-dotenv